[exp]drivers: input: ts: s2w: Add keypad S2W for Redmi 1S (armani)
*hack alert*

use the capacitive touch buttons on the armani for Sweep2Wake.
with the ft5x06 panel having a weird implementation of the keypad
layout, and firmware not exactly returning any touches other than
the buttons themselves (for y > 1280), use a simple algo to replicate
Sweep2Wake behaviour on the armani.

note: since the ft5x06 panel does wierd shiz, or my incompetence
with the algorithm, but this doesn't work sometimes (maybe the input
handler's borked, or input_sync() not being called 'correctly')
but just deal with it :stuck_out_tongue: